# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restobar Venecia Peruana",
        "description": "Sistema de restaurante en linea. Con control de inventario",
        "contact": {
            "email": "neisserrey@upeu.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://veneciaperuana.dev.com/api/",
            "description": "Documentación del Restobar Venecia Peruana"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Mostrar Categorias de productos",
                "operationId": "App\\Http\\Controllers\\Api\\CategoryController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de Categorias de productos. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Agregar un nuevo Categorias de productos.",
                "operationId": "store",
                "requestBody": {
                    "description": "Categorias de productos a ser creado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/categories/{categoryID}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Buscar Categorias de productos por ID",
                "description": "Retorna un Categorias de productos",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "categoryID",
                        "in": "path",
                        "description": "ID del Categorias de productos a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "Categorias de productos no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/categories/{categoryId}": {
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Actualizar una Categorias de productos existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID del Categorias de productos a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Categorias de productos a ser actualizada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Categorias de productos no encontrado"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Categorias de productos actualizado correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Elimina un Categorias de productos",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id del Categorias de productos a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Categorias de productos no encontrado"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Categorias de productos eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "name": {
                        "description": "name of the Category",
                        "type": "string"
                    },
                    "description": {
                        "description": "description of the Category",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "date when Category was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "date when Category was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}